refactor Height 클래스 생성
