+++ src/main/java/nextstep/ladder/model/Ladder.java
@@ -8,6 +8,8 @@ import java.util.stream.IntStream;
 public class Ladder {
 
     private static final String NOT_NULL_CREATE_MESSAGE = "사다리는 생성이 되어야 합니다.";
+    private static final String EMPTY_EXCEPTION_MESSAGE = "사다리의 높이는 1보다 작을 수 없습니다. (입력받은 사다리 높이:%d)";
+    private static final int LEAST_HEIGHT_SIZE = 1;
     private static final int START_INDEX = 0;
 
     private final List<Line> ladder;
@@ -29,9 +31,16 @@ public class Ladder {
     }
 
     private static List<Line> create(int participant, int height) {
-        return IntStream.range(START_INDEX, new Height(height).value())
+        validateHeightLength(height);
+        return IntStream.range(START_INDEX, height)
                 .mapToObj(i -> new Line(participant))
                 .collect(Collectors.toList());
+    }
+
+    private static void validateHeightLength(int value) {
+        if (value < LEAST_HEIGHT_SIZE) {
+            throw new NullPointerException(String.format(EMPTY_EXCEPTION_MESSAGE, value));
+        }
     }
 
     public List<Line> lines() {
